further restructure project added versioning to routes
